openpilot v0.9.9 release
date: 2025-04-05T09:09:42
master commit: 490c53e2dcb5e4c88e025bdb20a54a594d51377f

+++ tinygrad_repo/extra/gemm/simple_matvec.py
@@ -24,7 +24,7 @@ if __name__ == "__main__":
   for i in range(CNT):
     if i > 0 and getenv("RAND", 0) != 0:
       a, b = _rand(device)
-    c = a.matmul(b, acc_dtype=acc_dtype).realize()
+    c = a.matmul(b, dtype=acc_dtype).realize()
   nc = c.numpy()
   comp = a.numpy().astype(np.float32) @ b.numpy().astype(np.float32)
   np.testing.assert_allclose(nc, comp, atol=ATOL, rtol=RTOL)
